css change

+++ themes/wind/views/album.html.php
@@ -21,10 +21,15 @@ function showItemCover(v){
     //alert('g-item-cover-'+v)
     clearItemCovers();
     document.getElementById('g-item-panel-'+v).style.display='block';
-    //document.getElementById('<?='g-item-panel-'.$child->id?>').style.display='none'
+    //document.getElementById('<?//='g-item-panel-'.$child->id?>').style.display='none'
+}
+function showDefaultCover(){
+    clearItemCovers();
+    document.getElementById('g-item-panel-default').style.display='block';
 }
 $(function(){
     clearItemCovers();
+    showDefaultCover();
 })
 </script>
 <? endif ?>
@@ -36,12 +41,15 @@ $(function(){
 <div id="g-add-items">
     <? if ($theme->item->level < 2): ?>
     <ul id="album_header">
-        <li id="album_header_cpation">Albums</li>
+        <li id="album_header_cpation"><?= t("Albums") ?></li>
+        <? if (access::can("add", $item)): ?>
         <li id="album_header_new"><?= $theme->add_album_menu() ?></li>
+        <? endif ?>
     </ul>
     <? else: ?>
-      <?= $theme->add_photos_menu() ?>
-      
+        <? if (access::can("add", $item)): ?>
+            <?= $theme->add_photos_menu() ?>
+        <? endif ?>
     <? endif ?>
 </div>
 <ul id="g-album-grid" class="ui-helper-clearfix <?= ($theme->item->level < 2) ? "album_container" : "photo_container"?>">
@@ -56,9 +64,9 @@ $(function(){
     <a href="<?= $child->url() ?>">
       <? if ($child->has_thumb()): ?>
         <? if ($child->is_album()): ?>
-
+            <? // if is Album, no thumb img to show.?>
         <? elseif ($child->is_photo()): ?>
-        <?= $child->thumb_img(array("class" => "g-thumbnail"), 100) ?>
+            <?= $child->thumb_img(array("class" => "g-thumbnail ui-corner-all"), 100) ?>
         <? endif ?> 
       <? endif ?>
     </a>
@@ -72,25 +80,40 @@ $(function(){
   </li>
   <? endforeach ?>
 <? else: ?>
-  <? if ($user->admin || access::can("add", $item)): ?>
-  <? $addurl = url::site("uploader/index/$item->id") ?>
-  <li><?= t("There aren't any photos here yet! <a %attrs>Add some</a>.",
-            array("attrs" => html::mark_clean("href=\"$addurl\" class=\"g-dialog-link\""))) ?></li>
-  <? else: ?>
-  <li><?= t("There aren't any photos here yet!") ?></li>
-  <? endif; ?>
+<!--  <? // if ($user->admin || access::can("add", $item)): ?>
+  <? //$addurl = url::site("uploader/index/$item->id") ?>
+  <li><? //= t("There aren't any photos here yet! <a %attrs>Add some</a>.",
+          //  array("attrs" => html::mark_clean("href=\"$addurl\" class=\"g-dialog-link\""))) ?></li>
+  <? // else: ?>
+  <li><? //= t("There aren't any photos here yet!") ?></li>
+  <? // endif; ?>-->
+  
+  <?= t("There aren't any items yet!")?>
 <? endif; ?>
 </ul>
 <? if ($theme->item->level < 2): ?>
 <!-- CutGallery - ADDED -->
 <div id="g-item-panel">    
+       <ul id="g-item-panel-default" class="album_panel_parent">
+            <li id="g-item-cover-default" class="album_cover">
+                <div>
+                    <?//= $child->thumb_img(array("class" => "g-album-thumbnail ui-corner-all")) ?>
+                </div>
+            </li>
+            <li id="g-item-prop-default" class="album_prop">
+
+            </li>
+            <li id="g-item-comments-default" class="album_comments">
+                <span class="g-description"></span>
+            </li>
+        </ul>
        <? if (count($children)): ?>
           <? foreach ($children as $i => $child): ?>
                 <ul id="g-item-panel-<?= $child->id ?>" class="album_panel_parent">
                     <li id="g-item-cover-<?= $child->id ?>" class="album_cover">
                         <div>
                             <? if ($child->is_album()): ?>
-                                <a href='<?= $child->url() ?>'><?= $child->thumb_img(array("class" => "g-thumbnail")) ?></a>
+                                <a href='<?= $child->url() ?>'><?= $child->thumb_img(array("class" => "g-album-thumbnail ui-corner-all")) ?></a>
                             <? endif ?>
                         </div>
                     </li>
@@ -98,14 +121,16 @@ $(function(){
                         <span>Album Name: <?= html::purify($child->title) ?></span>
                         <span>Owner: <?= $child->owner_id ?></span>
                         <span>Pic Quantity:</span>
-                        <?= $theme->edit_album_menu($child) ?>
-                        <?= $theme->delete_album_menu($child) ?>
-                        <?= $theme->add_photos_menu($child) ?>
+                        <? if (access::can("add", $item)): ?>
+                            <?= $theme->edit_album_menu($child) ?>
+                            <?= $theme->delete_album_menu($child) ?>
+                            <?= $theme->add_photos_menu($child) ?>
+                        <? endif ?>
                     </li>
                     <li id="g-item-comments-<?= $child->id ?>" class="album_comments">
                         <span class="g-description"><?= nl2br(html::purify($child->description)) ?></span>
                     </li>
-                    </ul>
+                </ul>
         <? endforeach ?>
        <? endif ?>
 </div>
